RemoveComponentsOnTrigger, ToggleComponentsOnTrigger (#39639)
(cherry picked from commit d939b4dec795ccdc5d94578153d14716ed86f753)
